requirements.txt
APIDjango==1.3
arrow==1.2.1
asgiref==3.4.1
blessed==1.19.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.1
django-ckeditor==6.2.0
django-colorfield==0.6.3
Django-Deployment-Tools==0.5.0
django-filter==21.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-phone-verify==2.0.1
django-phonenumber-field==6.0.0
django-picklefield==3.0.1
django-q==1.3.9
django-qr-code==3.0.0
django-toggle-switch-widget==0.1.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-jsonapi==5.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.4.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
nexmo==2.5.2
packaging==21.3
pbr==5.8.0
phonenumbers==8.12.40
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
segno==1.4.1
six==1.16.0
sqlparse==0.4.2
twilio==7.4.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.11.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.13.3
